Example Part Update and refactor
