+++ example/lib/data/countries.dart
@@ -1,0 +1,43 @@
+import '../models/country.dart';
+
+const countries = [
+  Country(
+    flag: '🇦🇴',
+    name: 'Angola',
+    isoCode: 'AGO',
+    dialCode: '+244',
+    isAvailableMarket: true,
+    sectionLabel: 'AVAILABLE MARKETS',
+  ),
+  Country(
+    flag: '🇨🇩',
+    name: 'DR Congo',
+    isoCode: 'COD',
+    dialCode: '+243',
+    isAvailableMarket: true,
+  ),
+  Country(
+    flag: '🇬🇲',
+    name: 'Gambia',
+    isoCode: 'GMB',
+    dialCode: '+220',
+    isAvailableMarket: true,
+  ),
+  Country(
+    flag: '🇸🇱',
+    name: 'Sierra Leone',
+    isoCode: 'SLE',
+    dialCode: '+232',
+    isAvailableMarket: true,
+  ),
+  Country(
+    flag: '🇦🇫',
+    name: 'Afghanistan',
+    isoCode: 'AFG',
+    dialCode: '+93',
+    sectionLabel: 'ALL COUNTRIES',
+  ),
+  Country(flag: '🇦🇱', name: 'Albania', isoCode: 'ALB', dialCode: '+355'),
+  Country(flag: '🇩🇿', name: 'Algeria', isoCode: 'DZA', dialCode: '+213'),
+  Country(flag: '🇦🇷', name: 'Argentina', isoCode: 'ARG', dialCode: '+54'),
+];
